third_party/blink/web_tests/svg/dynamic-updates/SVGMaskElement-svgdom-height-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' property of the SVGMaskElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS maskElement.height.baseVal.valueAsString is "10%"
PASS maskElement.height.baseVal.valueAsString is "100%"
PASS successfullyParsed is true

TEST COMPLETE
